fix: globals to Micra namespace, partial config on run/start and static or instance on app config
